bricks: Fix a suble race in shmem::Fifo.
Since push first updates the ->next pointer in the current tail, a brief window
exists where head->next is not null even though head == tail. This might trick
pop() that's currently removing the last element in the queue into doing a
dropHead while in this window, causing head to overtake tail. If the reader can
then proceed to an empty() check quickly enough, it'll find that head != tail
(because head points at a node that *will become tail in the future*). This
could cause all sorts of bad behaviour, like reading non-existent items off the
queue or popping an empty queue causing items to be lost.