bricks: Extend Maybe and Union a bit.
Added fromMaybe and explcit operator bool to Maybe.

Union's conversion operator is now explicit which might break some code,
but it was a mistake to make it implicit. Furthermore, it now allows to do
static_cast internally, so something like double( v ) where v is
Union< int, double > is guaranteed to succeed. If conversion is impossible,
it will not compile.
Function convert is alias to conversion operator.
Furthermore, Union is now comparable, both to other union of *same* type
and to any of the possible member types.